+++ Zooper.Bee/Features/IWorkflowFeature.cs
@@ -7,15 +7,5 @@ namespace Zooper.Bee.Features;
 /// </summary>
 /// <typeparam name="TPayload">The type of the main workflow payload</typeparam>
 /// <typeparam name="TError">The type of the error</typeparam>
-public interface IWorkflowFeature<in TPayload, TError>
-{
-	/// <summary>
-	/// Gets the condition that determines if this feature should execute.
-	/// </summary>
-	Func<TPayload, bool>? Condition { get; }
-
-	/// <summary>
-	/// Whether this feature should merge back into the main workflow.
-	/// </summary>
-	bool ShouldMerge { get; }
-}
+[Obsolete("Use IRailwayFeature<TPayload, TError> instead. This interface will be removed in a future version.")]
+public interface IWorkflowFeature<in TPayload, TError> : IRailwayFeature<TPayload, TError>;