+++ Zooper.Bee/Features/Parallel/ParallelBuilder.cs
@@ -12,11 +12,11 @@ namespace Zooper.Bee.Features.Parallel;
 /// <typeparam name="TError">The type of the error result</typeparam>
 public sealed class ParallelBuilder<TRequest, TPayload, TSuccess, TError>
 {
-	private readonly WorkflowBuilder<TRequest, TPayload, TSuccess, TError> _workflow;
+	private readonly RailwayBuilder<TRequest, TPayload, TSuccess, TError> _workflow;
 	private readonly Parallel<TPayload, TError> _parallel;
 
 	internal ParallelBuilder(
-		WorkflowBuilder<TRequest, TPayload, TSuccess, TError> workflow,
+		RailwayBuilder<TRequest, TPayload, TSuccess, TError> workflow,
 		Parallel<TPayload, TError> parallel)
 	{
 		_workflow = workflow;